refactor: rename authStore to auth-store for consistent kebab-case naming
- Rename src/stores/authStore.ts to src/stores/auth-store.ts
- Update import in src/main.tsx to use new kebab-case path
